Initial functionality as desired: switch and move bangs list the active tasks.
Moved !task and !movetotask into a specific tabMoverPlugin.
Used the searchbarPlugins.addResult() to add task items
