+++ js/default.js
@@ -161,6 +161,7 @@ require('searchbar/instantAnswerPlugin.js').initialize()
 require('searchbar/openTabsPlugin.js').initialize()
 require('searchbar/bangsPlugin.js').initialize()
 require('searchbar/customBangs.js').initialize()
+require('searchbar/tabMoverPlugin.js').initialize()
 require('searchbar/searchSuggestionsPlugin.js').initialize()
 require('searchbar/placeSuggestionsPlugin.js').initialize()
 require('searchbar/updateNotifications.js').initialize()
