+++ js/searchbar/customBangs.js
@@ -90,32 +90,32 @@ function initialize () {
     )
   }
 
-  bangsPlugin.registerCustomBang({
-    phrase: '!task',
-    snippet: l('switchToTask'),
-    isAction: false,
-    fn: function (text) {
-    /* disabled in focus mode */
-      if (focusMode.enabled()) {
-        focusMode.warn()
-        return
-      }
+  // bangsPlugin.registerCustomBang({
+  //   phrase: '!task',
+  //   snippet: l('switchToTask'),
+  //   isAction: false,
+  //   fn: function (text) {
+  //   /* disabled in focus mode */
+  //     if (focusMode.enabled()) {
+  //       focusMode.warn()
+  //       return
+  //     }
 
-      text = text.toLowerCase()
+  //     text = text.toLowerCase()
 
-      // no task was specified, show all of the tasks
-      if (!text) {
-        taskOverlay.show()
-        return
-      }
+  //     // no task was specified, show all of the tasks
+  //     if (!text) {
+  //       taskOverlay.show()
+  //       return
+  //     }
 
-      var task = getTaskByNameOrNumber(text)
+  //     var task = getTaskByNameOrNumber(text)
 
-      if (task) {
-        browserUI.switchToTask(task.id)
-      }
-    }
-  })
+  //     if (task) {
+  //       browserUI.switchToTask(task.id)
+  //     }
+  //   }
+  // })
 
   bangsPlugin.registerCustomBang({
     phrase: '!newtask',
@@ -139,48 +139,48 @@ function initialize () {
     }
   })
 
-  bangsPlugin.registerCustomBang({
-    phrase: '!movetotask',
-    snippet: l('moveToTask'),
-    isAction: false,
-    fn: function (text) {
-    /* disabled in focus mode */
-      if (focusMode.enabled()) {
-        focusMode.warn()
-        return
-      }
+  // bangsPlugin.registerCustomBang({
+  //   phrase: '!movetotask',
+  //   snippet: l('moveToTask'),
+  //   isAction: false,
+  //   fn: function (text) {
+  //   /* disabled in focus mode */
+  //     if (focusMode.enabled()) {
+  //       focusMode.warn()
+  //       return
+  //     }
 
-      // remove the tab from the current task
+  //     // remove the tab from the current task
 
-      var currentTab = tabs.get(tabs.getSelected())
-      tabs.destroy(currentTab.id)
+  //     var currentTab = tabs.get(tabs.getSelected())
+  //     tabs.destroy(currentTab.id)
 
-      // make sure the task has at least one tab in it
-      if (tabs.count() === 0) {
-        tabs.add()
-      }
+  //     // make sure the task has at least one tab in it
+  //     if (tabs.count() === 0) {
+  //       tabs.add()
+  //     }
 
-      var newTask = getTaskByNameOrNumber(text)
+  //     var newTask = getTaskByNameOrNumber(text)
 
-      if (newTask) {
-        newTask.tabs.add(currentTab, { atEnd: true })
-      } else {
-      // create a new task with the given name
-        newTask = tasks.get(tasks.add(undefined, tasks.getIndex(tasks.getSelected().id) + 1))
-        newTask.name = text
+  //     if (newTask) {
+  //       newTask.tabs.add(currentTab, { atEnd: true })
+  //     } else {
+  //     // create a new task with the given name
+  //       newTask = tasks.get(tasks.add(undefined, tasks.getIndex(tasks.getSelected().id) + 1))
+  //       newTask.name = text
 
-        newTask.tabs.add(currentTab)
-      }
+  //       newTask.tabs.add(currentTab)
+  //     }
 
-      browserUI.switchToTask(newTask.id)
-      browserUI.switchToTab(currentTab.id)
-      taskOverlay.show()
+  //     browserUI.switchToTask(newTask.id)
+  //     browserUI.switchToTab(currentTab.id)
+  //     taskOverlay.show()
 
-      setTimeout(function () {
-        taskOverlay.hide()
-      }, 600)
-    }
-  })
+  //     setTimeout(function () {
+  //       taskOverlay.hide()
+  //     }, 600)
+  //   }
+  // })
 
   bangsPlugin.registerCustomBang({
     phrase: '!closetask',
